Fix a UI bug in firefox
Firefox doesn't count the dropdown menu as part of the <select> element, so when a user tries to click on one of the options, it triggers onmouseout and blurs the menu. It's possible to use the arrow keys but is still difficult to figure out. I just removed that, since there's no good, readable way of removing it for firefox only.